chore(audit): complete non-cookie production audit hardening
- implement remaining backend/frontend audit fixes; `OPEN` items reduced to 0 in `docs/AUDIT.md`
- harden security defaults (JWT library migration, rate limits, CORS/headers, auth/health protections)
- improve websocket resilience and reconnect/read-state behavior
- add DB integrity migrations (`ON DELETE CASCADE`, PK/FK bigint) and SQLAlchemy 2.0 model alignment
- fix frontend a11y/UX issues (aria labels, modal focus trap, escape handling, routing links, logging behavior)
- expand targeted tests and run verification (`frontend build/lint/typecheck`, `backend-verify SKIP_DB_BOOTSTRAP=1`)
- leave cookie auth migration as planned follow-up (`COOKIE_AUTH_IMPLEMENTATION_PLAN.md`) and tooling strictness as separate pass

+++ backend/app/models/user.py
@@ -1,15 +1,16 @@
 from datetime import UTC, datetime
 
-from app.core.database import Base
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import BigInteger, DateTime, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
+
+from app.core.database import Base
 
 
 class User(Base):
     __tablename__ = "users"
 
     # Columns with proper type hints for mypy
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True)
     username: Mapped[str] = mapped_column(
         String, unique=True, index=True, nullable=False
     )
